move and test toHumanReadable
